Refactor: 영화선택 부분 구조 분리
