+++ Lotte-Cinema/src/hooks/query/index.ts
@@ -1,3 +1,4 @@
 import { useMovieListQuery } from './useMovieListQuery';
+import { useMovieTimeQuery } from './useMovieTime';
 
-export { useMovieListQuery };
+export { useMovieListQuery, useMovieTimeQuery };
