+++ Lotte-Cinema/src/types/timeSelect/index.ts
@@ -1,0 +1,6 @@
+export type MovieInfoBarProps = {
+  locs: string[];
+  onDelete: (loc: string) => void;
+  setSelectTitle: (value: string) => void;
+  selectTitle: string;
+};
